# makefiles: Use simple expansion for widely used variables