### Contribution description
This PR changes some Makefile variable assignments to simple expansion, so we don't have to shell out multiple times when not required on widely used variables.

One can see all the calls to `basename`, for instance, by doing `make SHELL='/bin/sh -x' 2>&1 | less`.

### Test results 
I built **50 times** a couple of applications, for different boards, and measured with `time` each build. These are the averages:
#### samr21-xpro:`examples/gcoap`
- On master:
  - Real    3.85 s
  - User    2.95 s
  - System  0.96 s

- With this PR:
  - Real    3.02 s
  - User    2.38 s
  - System  0.66 s

**Time decreases in ~21.5%**.

#### native:`examples/gnrc_networking`
- On master:
  - Real    2.28 s
  - User    1.63 s
  - System  0.70 s

- With this PR:
  - Real    1.91 s
  - User    1.33 s
  - System  0.60 s

**Time decreases in ~16.2%**.

### Testing procedure
- Green Murdock
- You can measure time with something like:
```bash
BOARD=samr21-xpro make -C examples/gcoap clean
time -p BOARD=samr21-xpro make -C examples/gcoap all
```

### Issues/PRs references
None