@cgundogan 2020-02-05
let's see how murdock reacts to this optimization