@kaspar030 2020-02-05
Nice one!
I can confirm the speed increase:

This PR:

```
[kaspar@ng riot.tmp/examples/gcoap (pr13292)]$ hyperfine -w1 -m20 "BOARD=samr21-xpro make -j8"
Benchmark #1: BOARD=samr21-xpro make -j8                                                                                         
  Time (mean ± σ):      1.032 s ±  0.028 s    [User: 1.543 s, System: 0.341 s]
  Range (min … max):    1.009 s …  1.088 s    20 runs
```

master:

```
[kaspar@ng riot.tmp/examples/gcoap ((769209351a...))]$ hyperfine -w1 -m20 "BOARD=samr21-xpro make -j8"
Benchmark #1: BOARD=samr21-xpro make -j8
  Time (mean ± σ):      1.224 s ±  0.073 s    [User: 1.984 s, System: 0.447 s]
  Range (min … max):    1.146 s …  1.395 s    20 runs
```

1.032 vs 1.224 sec, on a Thinkpad with i5-8250U, Linux, performance governor.