@fjmolinas 2020-02-05
Nice catch! @leandrolanzieri are these variables **allways**? if not maybe you could use `memoized`?