@kaspar030 2020-02-05
> @kaspar030 murdock does not seem to be affected by this speed increase. Any ideas why?

Murdock builds *massively* parrallel (e.g., N times "make -j4" for N cores), in order to keep the CPUs busy. I'd assume that the fix in this PR doesn't increase the throughput much in that setting, CPU's waiting to be fed are just scheduled to compile something else.

But it might actually do speedup some things. Murdock build times are quite variable, but if you compare the average build times of the workers "riotbuild" and "mobi9..." (both have 64 cores, the former Epycs, the latter some older Opterons IIRC),  they were significantly faster in this PR vs. the previous two successfully builds.

this PR: https://ci.riot-os.org/RIOT-OS/RIOT/13292/fdff969695bdb3ef4b67763da6766b93aa7c498e/output.html
Previous: https://ci.riot-os.org/RIOT-OS/RIOT/13290/5391e50bb7721895d3406ed1dde7352113b74190/output.html
Another: https://ci.riot-os.org/RIOT-OS/RIOT/12969/c4e0ed4b7991683f6dfd25fd2905dcc43e0d371e/output.html

(scroll all the way down to see the worker stats)