@kaspar030 2020-02-05
> @leandrolanzieri are these variables **allways**? if not maybe you could use `memoized`?

That sounds interesting, but I have no clue what it means. Do you have a pointer?