@leandrolanzieri 2020-02-05
> Nice catch! @leandrolanzieri are these variables **allways**? 

I think in the case of `MODULE` yes, as it is used for building the targets in `Makfile.base`.

> if not maybe you could use `memoized`?

I can measure using  `memoized` for the other variables.