@leandrolanzieri 2020-02-05
> Nice catch! @leandrolanzieri are these variables **allways** used?

Also, I see now `OBJCOPY` is evaluated right after so it's expanded.